test: rename folder
